[libc++][WIP] Move the libc++ test format to Lit
This allows the generally-useful parts of the test format to be
shipped alongside Lit, which would make it usable for third-party
developers as well. This came up at C++Now in the context of the
Beman project where we'd like to set up a non-LLVM project that can
use the same testing framework as libc++.

With the test format moved to Lit, the format would be available
after installing Lit with pip, which is really convenient.
