+++ llvm/utils/lit/lit/formats/__init__.py
@@ -6,3 +6,4 @@ from lit.formats.base import (  # noqa: F401
 
 from lit.formats.googletest import GoogleTest  # noqa: F401
 from lit.formats.shtest import ShTest  # noqa: F401
+from lit.formats.standardlibrarytest import StandardLibraryTest  # noqa: F401
